refactor: change approach to get way
