+++ services/functionsForUsing/transfer.js
@@ -1,7 +1,11 @@
-import { checkMove } from '../checkMove.js';
+import { checkMove, getCurrentPos } from '../checkMove.js';
+import { getListOfWays } from '../handlers/getListOfWays.js';
 
 export default function transfer (arr, gameDOM) {
   return (from, to) => {
-    console.log(checkMove(from, to, arr));
+    const way = checkMove(from, to, arr);
+    if (way.toPos && way.spacePos) {
+      const listOfWays = getListOfWays(way, getCurrentPos(from, arr));
+    }
   }
 }
