+++ services/handlers/getListOfWays.js
@@ -1,0 +1,7 @@
+function getListOfWays (way, chosenPos) {
+  console.log(chosenPos);
+  console.log(way.toPos);
+  console.log(way.spacePos);
+}
+
+export { getListOfWays };
